fix(api): fixes for tesla 20.49 API changes
With the 20.49 firmware update by Tesla, the endpoints that previously
were open are no longer available without authentication. Unfortuantely,
the mechanism used for authentication is not HTTP basic, rather it's
cookie based. This change brings in an additional container that I
created called cookieproxy, that can be used to proxy the cookie based
requests through to the Powerwall gateway.

In contrast to previous versions, this requires no compilation of
containers on the local machine and no modifications to the more common
containers of influxdb, grafana, and telegraf. There is a small
customization that is required to store the password of your Powerwall
in the `.env.cookieproxy` file, but otherwise it should be plug and play
with the previous versions of mihailescu2m/powerwall_monitor.

See mihailescu2m/powerwall_monitor#14

DCO 1.1 Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>